Create new stage for NMODL related tests
This is MR is needed for https://github.com/BlueBrain/nmodl/pull/1037.
It adds a new stage that includes only selected tests that have CoreNEURON runs and cover all neurodamus-models and some important circuit features.
Those tests at the moment are:
- quick-hip-delayconn
- sscx-v7-plasticity
- sonataconf-quick-scx-multi-circuit
- sonataconf-quick-thalamus
- mousify
- scx-v5-gapjunctions

All the other tests are disabled if `NMODL_RUN` is set.
This is tested in https://github.com/BlueBrain/nmodl/pull/1037