[BBPBGLIB-1090]Migrate tests with BlueConfig files to SONATA tests
It is planned to remove the support for BlueConfig files in neurodamus, therefore all the CI tests should be using SONATA config files. This MR either converts the current BlueConfig files to SONATA config files, or replaces the BlueConfig tests with recent SONATA simulations. Tests using neocortex, hippocampus and thalamus models are based on the recent simulations recommended by scientists. The mousify test is converted from the BlueConfig test which uses a sonata circuit.