Avoid creating new reference files every compiler version change
The new spack deployment changes the oneapi compiler version from 2022 to 2023.
CI was failing because it didn't find the exact version.

This fixes the issue by checking older versions of the same compiler.
Whenever the result change with a different compiler version, then we would need to update the references.