Add simulation with etypes prepared to drop axon
### Context

For NGV and future simulations we want to inspect the axon before replacing it with a minimal stub.

Such replacement must be backwards compatible, so it is particularly important to not rename public accessible structures

We build on https://bbpgitlab.epfl.ch/molsys/multiscale_run/-/merge_requests/43 and avoid changing variable names.

However Neuron is very sensitive to recreate sections with the same name in multiple times in procedures. Moreover, there is a chance those sections are already being referenced somewhere.

### In this PR

We changed the implementation so that 

1. We insert AXON mechanisms and define their properties only in the very first two sections.  


1. we _REUSE_ the first two sections of the axon for the stub, only modifying their properties.


This mini simulation count 15 cells with 2 different ETypes

Preliminary results are promising